Move the next level out of ast
