+++ edn/_ast.py
@@ -1,9 +1,6 @@
-import datetime
 from functools import partial
 import os
-import uuid
 
-import iso8601
 from parsley import makeGrammar
 from terml.nodes import termMaker as t
 
@@ -18,15 +15,6 @@ Symbol = t.Symbol
 TaggedValue = t.TaggedValue
 Vector = t.Vector
 
-
-INST = t.Symbol('inst')
-UUID = t.Symbol('uuid')
-
-
-BUILTIN_READ_HANDLERS = {
-    INST: iso8601.parse_date,
-    UUID: uuid.UUID,
-}
 
 
 _edn_grammar_file = os.path.join(os.path.dirname(__file__), 'edn.parsley')
@@ -52,37 +40,11 @@ def parse(string):
     return edn(string).edn()
 
 
-def loads(string, handlers=None):
-    if handlers is None:
-        handlers = BUILTIN_READ_HANDLERS
-    return parse(string)
-
-
 def _dump_bool(obj):
     if obj:
         return 'true'
     else:
         return 'false'
-
-
-# XXX: This is a poor way of doing type-based dispatch.  Some thoughts:
-# - are we sure that we _always_ want to do type-based dispatch?  the
-#   most flexible way to do this is to have arbritrary predicates, or
-#   a list of functions that return some marker value if they don't
-#   know what to do
-# - clojure does typed-based dispatch, but it has multimethods
-# - if we did type-based dispatch, we could use Python's ABC, I guess.
-# - does it make sense to allow callers to override the behaviour of
-#   the standard types, e.g. to encode bools differently
-# - does it even make sense to allow callers to overwrite the
-#   built-in write handlers?  the current API requires you specify them.
-# - perhaps a global registry wouldn't be such a bad thing?
-
-
-def tagger(tag, function):
-    def wrapped(*args, **kwargs):
-        return t.TaggedValue(tag, function(*args, **kwargs))
-    return wrapped
 
 
 def _wrap(start, end, middle):
@@ -164,21 +126,3 @@ def unparse(obj):
     if build:
         return build(builder)
     return builder.leafData(obj)(obj)
-
-
-DEFAULT_WRITE_HANDLERS = [
-    (datetime.datetime, tagger(INST, lambda x: x.isoformat())),
-    (uuid.UUID, tagger(UUID, str)),
-]
-
-
-def dumps(obj, write_handlers=None):
-    if write_handlers is None:
-        write_handlers = DEFAULT_WRITE_HANDLERS
-
-    for base_type, function in write_handlers:
-        if isinstance(obj, base_type):
-            obj = function(obj)
-            break
-
-    return unparse(obj)
